Split FenceLink out to its own file for testing
